Add a PLL mechanism to lock the master on the DC ref clock (#429)
OS/Time: make seconds_f a double to avoid long-run precision loss
marvin: 1 Hz PLL/DC telemetry with smoothed drift

+++ lib/include/kickcat/OS/Time.h
@@ -7,7 +7,8 @@
 namespace kickcat
 {
     using namespace std::chrono;
-    using seconds_f = std::chrono::duration<float>;
+    // double, not float despite the name: a float second count loses sub-ms resolution after ~30 min.
+    using seconds_f = std::chrono::duration<double>;
 
     void sleep(nanoseconds ns);
